Remove cyclic references
