+++ packages/database/src/exp/Database.ts
@@ -18,9 +18,8 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { _FirebaseService, _getProvider, FirebaseApp } from '@firebase/app-exp';
 import { Reference } from '../api/Reference';
-import { repoManagerDatabaseFromApp } from '../core/RepoManager';
 import { FirebaseAuthInternalName } from '@firebase/auth-interop-types';
-import { Database } from '../api/Database';
+import { Database, repoManagerDatabaseFromApp } from '../api/Database';
 import { Provider } from '@firebase/component';
 
 /**
